Move identifier validation to types module
