+++ src/language/types.rs
@@ -1,5 +1,7 @@
 // Abstract Syntax Trees for the Technique language
 
+use regex::Regex;
+
 #[derive(Eq, Debug, PartialEq)]
 pub struct Technique<'i> {
     pub version: u8,
@@ -17,6 +19,13 @@ impl Default for Technique<'_> {
             template: None,
         }
     }
+}
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum ValidationError {
+    ZeroLengthToken,
+    InvalidIdentifier,
+    InvalidForma,
 }
 
 #[derive(Eq, Debug, PartialEq)]
@@ -48,9 +57,43 @@ pub struct Attribute<'i> {
     pub name: &'i str,
 }
 
+// HERE ALSO ... validating identifiers should be here!!!
+
+pub fn validate_identifier(input: &str) -> Result<&str, ValidationError> {
+    if input.len() == 0 {
+        return Err(ValidationError::ZeroLengthToken);
+    }
+
+    let re = Regex::new(r"^[a-z][a-z0-9_]*$").unwrap();
+    if re.is_match(input) {
+        Ok(input)
+    } else {
+        Err(ValidationError::InvalidIdentifier)
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[test]
+    fn check_identifier_rules() {
+        assert_eq!(validate_identifier("a"), Ok("a"));
+        assert_eq!(validate_identifier("ab"), Ok("ab"));
+        assert_eq!(validate_identifier("johnny5"), Ok("johnny5"));
+        assert_eq!(
+            validate_identifier("Pizza"),
+            Err(ValidationError::InvalidIdentifier)
+        );
+        assert_eq!(
+            validate_identifier("pizZa"),
+            Err(ValidationError::InvalidIdentifier)
+        );
+        assert!(validate_identifier("0trust").is_err());
+        assert_eq!(validate_identifier("make_dinner"), Ok("make_dinner"));
+        assert!(validate_identifier("MakeDinner").is_err());
+        assert!(validate_identifier("make-dinner").is_err());
+    }
 
     fn maker<'i>() -> Technique<'i> {
         let t1 = Technique {
